armsr: armv8: update configs based on 2e3ea6d206
Initially, I added configs based on the Git diff extracted from history,
but that was not getting all the configs because before armsr, there was
armvirt.

All configs added here are based on a diff from 2e3ea6d206. Using the
SystemReady-Certified platforms plus ones pending to be certified.

Tested on qemu.

Signed-off-by: Javier Tia <javier.tia@linaro.org>